% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_points_to_polygons.R
\name{cells_to_polygons}
\alias{cells_to_polygons}
\title{Convert points for an entire reef that have been allocated site IDs into site polygons.}
\usage{
cells_to_polygons(
  clustered_points,
  site_id_col = "site_id",
  reef_cols_to_keep = c("clustering_time", "UNIQUE_ID")
)
}
\arguments{
\item{clustered_points}{data.frame. Contains a row for each point and cluster allocations.}

\item{site_id_col}{character or integer. Column containing site allocations for points.
Default = "site_id".}

\item{reef_cols_to_keep}{character vector. Columns containing unique values per reef that
can be allocated to the collated site polygons to conserve IDs and other reef level
information. Examples include reef level mean depth, or reef distance to coastline.
Default = c("clustering_time", "UNIQUE_ID").}
}
\value{
sf data.frame containing site polygons created from points using allocated
\code{site_id_col} values.
}
\description{
From a dataframe of points and point data, this function clusters points into
sites (within each habitat type), using a user defined clustering method that considers
both geographical data and additional extracted data (such as depth). \strong{cells}_to_polygons
refers to points that have been extracted using H3 methods. This method of polygon
creation uses H3 methods to polygonise covered areas.
}
