% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_reef_pixels.R
\name{cluster_reef_pixels}
\alias{cluster_reef_pixels}
\title{Cluster pixels into sites based on geographical attributes and additional extracted pixel values.}
\usage{
cluster_reef_pixels(
  pixels,
  habitat_col = "habitat",
  x_col = "X",
  y_col = "Y",
  additional_variable_cols = c("depth"),
  reef_id_col = "UNIQUE_ID",
  habitat_clustering_function = constrained_hclust_mst,
  clustering_function_args = list()
)
}
\arguments{
\item{pixels}{data.frame. Contains a row for each pixel and values for each variable to
be input into \code{habitat_clustering_function}.}

\item{habitat_col}{character or integer. Column containing the categorical habitat allocations
of pixels. Default = "habitat".}

\item{x_col}{character or integer. Column containing the continuous geographical X values
for pixels. Default = "X".}

\item{y_col}{character or integer. Column containing the continuous geographical Y values
for pixels. Default = "Y".}

\item{additional_variable_cols}{character vector. Vector containing column names for columns
with additional clustering variable values. This vector can have length > 1. Values will
be standardised and output as \verb{*variable*_standard}. Default = c("Depth").}

\item{reef_id_col}{character or integer. Column containing the single unique reef ID to
be assigned as an identifier before outputting cluster allocations. Default = "UNIQUE_ID".}

\item{habitat_clustering_function}{function. A function that takes a dataframe of pixels
for a single habitat type as the first argument and additional arguments, and returns
a dataframe of input pixels with an additional column containing assigned \code{site_id} values.
Additional arguments must include \code{px_per_cluster}, \code{habitat_col}, \code{x_col}, \code{y_col}
and \code{additional_variable_cols}.
Default options available include \code{reef_skater} and \code{constrained_hclust}.}

\item{px_per_cluster}{integer. The number of pixels to be allocated into each cluster using
\code{habitat_clustering_function}. If \code{habitat_clustering_function} defines k-clusters, this
must be \code{round(nrow(x) / px_per_cluster)}. Default = 200 pixels.}
}
\value{
data.frame containing pixels that have a site ID allocated in \code{site_id} column.
Output data includes clustering execution time.
}
\description{
From a dataframe of pixels and pixel data, this function clusters pixels into
sites (within each habitat type), using a user defined clustering method that considers
both geographical data and additional extracted data (such as depth).
}
