% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_reef_points.R
\name{cluster_reef_points}
\alias{cluster_reef_points}
\title{Cluster points into sites based on geographical attributes and additional extracted point values.}
\usage{
cluster_reef_points(
  points,
  habitat_col = "habitat",
  x_col = "X",
  y_col = "Y",
  additional_variable_cols = c("depth"),
  reef_id_col = "UNIQUE_ID",
  clustering_function_args = list()
  habitat_clustering_function = reef_skater_fast,
)
}
\arguments{
\item{points}{data.frame. Contains a row for each point and values for each variable to
be input into \code{habitat_clustering_function}.}

\item{habitat_col}{character or integer. Column containing the categorical habitat allocations
of points. Default = "habitat".}

\item{x_col}{character or integer. Column containing the continuous geographical X values
for points. Default = "X".}

\item{y_col}{character or integer. Column containing the continuous geographical Y values
for points. Default = "Y".}

\item{additional_variable_cols}{character vector. Vector containing column names for columns
with additional clustering variable values. This vector can have length > 1. Values will
be standardised and output as \verb{*variable*_standard}. Default = c("Depth").}

\item{reef_id_col}{character or integer. Column containing the single unique reef ID to
be assigned as an identifier before outputting cluster allocations. Default = "UNIQUE_ID".}

\item{habitat_clustering_function}{function. A function that takes a dataframe of points
for a single habitat type as the first argument and additional arguments, and returns
a dataframe of input points with an additional column containing assigned \code{site_id} values.
Additional arguments must include \code{px_per_cluster}, \code{habitat_col}, \code{x_col}, \code{y_col}
and \code{additional_variable_cols}.
Default options available include \code{reef_skater} and \code{constrained_hclust}.}
}
\value{
data.frame containing points that have a site ID allocated in \code{site_id} column.
Output data includes clustering execution time.
}
\description{
From a dataframe of points and point data, this function clusters points into
sites (within each habitat type), using a user defined clustering method that considers
both geographical data and additional extracted data (such as depth).
}
