% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_postprocessing.R
\name{clustered_pixels_to_polygons}
\alias{clustered_pixels_to_polygons}
\title{Convert pixels for an entire reef that have been allocated site IDs into site polygons.}
\usage{
clustered_pixels_to_polygons(
  clustered_pixels,
  site_id_col = "site_id",
  reef_cols_to_keep = c("clustering_time", "UNIQUE_ID")
)
}
\arguments{
\item{clustered_pixels}{data.frame. Contains a row for each pixel and cluster allocations.}

\item{site_id_col}{character or integer. Column containing site allocations for pixels.
Default = "site_id".}

\item{reef_cols_to_keep}{character vector. Columns containing unique values per reef that
can be allocated to the collated site polygons to conserve IDs and other reef level
information. Examples include reef level mean depth, or reef distance to coastline.
Default = c("clustering_time", "UNIQUE_ID").}
}
\value{
sf data.frame containing site polygons created from pixels using allocated
\code{site_id_col} values.
}
\description{
From a dataframe of pixels and pixel data, this function clusters pixels into
sites (within each habitat type), using a user defined clustering method that considers
both geographical data and additional extracted data (such as depth).
}
