% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_hclust.R
\name{constrained_hclust_mst}
\alias{constrained_hclust_mst}
\title{Default habitat clustering function using adespatial::constr.hclust.}
\usage{
constrained_hclust_mst(
  points,
  distance_alpha = 0.5,
  n_points = 204,
  x_col = "X_standard",
  y_col = "Y_standard",
  interpolation_threshold = 30000,
  ...
)
}
\arguments{
\item{points}{data.frame. Contains values for X and Y coordinates, as well as
\code{additional_variable_cols}.}

\item{distance_alpha}{float numeric. Weighting applied to the additional variable
distance values when creating the distance matrix for clustering. This argument
is not included in \code{...} for discoverability.}

\item{n_points}{numeric. Desired number of points per cluster. Default = 204 (
which equates to site sizes of 62,500m^2 when using H3 cells of resolution 12).}

\item{x_col}{character. Name of column holding X coordinate values. Default =
"X_standard".}

\item{y_col}{character. Name of column holding Y coordinates. Default = "Y_standard".}

\item{interpolation_threshold}{numeric. Threshold from where to sample random
points and interpolate clusters for remaining points. This value should be
scaled with reef area for larger reefs. Default value is 30,000, setting a higher
threshold may result in long computation times and high RAM usage.}

\item{...}{additional arguments. Additional arguments can be used here and will
be passed onto \code{prepare_mst_edges()} and \code{constrained_hclust()} functions.
These arguments must be named. \code{distance_alpha} argument is not included in
these additional arguments. For information on arguments available in these
functions and default values when arguments are not used, see \code{prepare_mst_edges()}
and \code{constrained_hclust()}.}
}
\value{
data.frame of points with allocated site_ids based on cluster outputs
from \code{constrained_hclust()} using \code{prepare_mst_edges} to create a minimum
spanning tree for input. \code{site_id} values are a combination of the \code{id_col}
value, \code{habitat_col} value and the cluster allocation.
}
\description{
Take a dataframe of points containing \code{additional_variable_cols}
values, create a minimum spanning tree using \code{prepare_mst_edges()} and then
cluster points using point data and edges with \code{constrained_hclust()}.
Any additional arguments for \code{prepare_mst_edges()} or \code{constrained_hclust()}
(excluding \code{distance_alpha}) can be included.
}
