% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pixel_points.R
\name{extract_pixel_points}
\alias{extract_pixel_points}
\title{Extract pixel data from selected habitats over a target reef.}
\usage{
extract_pixel_points(
  reef_polygon,
  habitat_raster,
  add_var_raster,
  habitat_categories,
  hex_resolution = 12,
  unit = "km2",
  additional_variable_name = "depth",
  output_epsg = 3112,
  resample_method = "bilinear"
)
}
\arguments{
\item{reef_polygon}{sf_object. sf object containing the target reef polygons for data
coverage.}

\item{habitat_raster}{SpatRaster. Terra raster object containing categorical values for
habitats covering the target reef area.}

\item{add_var_raster}{SpatRaster. Terra raster object containing an additional variable
to extract for all selected habitat pixels covering the target reef area. Can be continuous
or categorical, however the raster will be resampled and the method should be changed for
categorical data.}

\item{habitat_categories}{character. Vector containing the habitat categories to select from
\code{habitat_raster}.}

\item{hex_resolution}{integer. Selected H3 resolution of hexagons for pixel representation.
Default = 12.}

\item{unit}{character. Unit used by H3 functions to calculate the area of hexagons.
Default = "km2".}

\item{additional_variable_name}{character. Name to assign the extracted data from
\code{add_var_raster}. Default = "depth".}

\item{output_epsg}{integer. EPSG code used for outputting pixels and extracted data.
Default = 3112.}

\item{resample_method}{Method used to resample \code{add_var_raster} before extracting pixel
data. Default = "bilinear", method should be changed for categorical data.
See \code{\link[terra:disaggregate]{terra::disagg()}} for more details.}
}
\value{
data.frame containing \code{habitat_raster} pixels covering \code{reef_polygon}
for selected habitats in \code{habitat_categories}, alongside extracted data from \code{add_var_raster}.
}
\description{
Extract data from the centroids of pixels that overlap the target \code{reef_polyon}.
Extraction is performed for two raster layers, one \code{habitat_raster} must contain habitat
categories that form the basis of the pixel centroids, the other \code{add_var_raster} is
additional desired data for clustering at later points in the workflow.
}
