% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_point_pixels.R
\name{extract_point_pixels}
\alias{extract_point_pixels}
\title{Extract pixel data from selected habitats over a target reef.}
\usage{
extract_point_pixels(
  reef_polygon,
  habitat_raster,
  add_var_raster,
  habitat_categories,
  additional_variable_name = "depth",
  output_epsg = 3112,
  interpolation = TRUE
)
}
\arguments{
\item{reef_polygon}{sf_object. sf object containing the target reef polygons for data
coverage.}

\item{habitat_raster}{SpatRaster. Terra raster object containing categorical values for
habitats covering the target reef area.}

\item{add_var_raster}{SpatRaster. Terra raster object containing an additional variable
to extract for all selected habitat pixels covering the target reef area. Can be continuous
or categorical, however the raster will be resampled and the method should be changed for
categorical data.}

\item{habitat_categories}{character. Vector containing the habitat categories to select from
\code{habitat_raster}.}

\item{additional_variable_name}{character. Name to assign the extracted data from
\code{add_var_raster}. Default = "depth".}

\item{output_epsg}{integer. EPSG code used for outputting pixels and extracted data.
Default = 3112.}

\item{interpolation}{bool. Option to interpolate missing values after extracting
data from \code{add_var_raster} using nearest neighbour interpolation.}
}
\value{
data.frame containing points for pixels from \code{habitat_raster} covering \code{reef_polygon}
for selected habitats in \code{habitat_categories}, alongside extracted data from \code{add_var_raster}.
}
\description{
Extract data from the centroids of pixels that overlap the target \code{reef_polyon}.
Extraction is performed for two raster layers, one \code{habitat_raster} must contain habitat
categories that form the basis of the pixel centroids, the other \code{add_var_raster} is
additional desired data for clustering at later points in the workflow. Extrction method
uses habitat_raster pixels, meaning points have a minimum distance apart of res(habitat_raster) / 2.
}
