% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_raster.R
\name{generate_raster_template}
\alias{generate_raster_template}
\title{Generate a raster template based on sf \code{vector_data} at a specific \code{pixel_resolution}.}
\usage{
generate_raster_template(vector_data, pixel_resolution = NA)
}
\arguments{
\item{vector_data}{sf data.frame. Vector data to generate a raster template object.}

\item{pixel_resolution}{numeric. Resolution of raster cells to use in template
creation. Default is NA. If pixel_resolution is NA then a default resolution
of 0.001 degrees or 100m will be used depending on the unit of \code{vector_data}
coordinate reference system.}
}
\value{
template raster object based on vector_data extent.
}
\description{
Generate a raster template based on sf \code{vector_data} at a specific \code{pixel_resolution}.
}
