% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_points_to_polygons.R
\name{hex_to_polygons}
\alias{hex_to_polygons}
\title{Helper function to convert points in a single site into an sf poylgon.}
\usage{
hex_to_polygons(x, h3_id_col = "id", site_id_col = "site_id")
}
\arguments{
\item{x}{data.frame. Contains point values for the target site.}

\item{h3_id_col}{character or integer. Column containing the H3 ID values for each point
to be collated into the polygon. Default = "id".}

\item{site_id_col}{charcater vector. Column containing the unique site ID for the target
site contained in \code{x}. Default = "site_id".}
}
\value{
sf data.frame containing site polygons created from points using allocated
\code{site_id_col} values.
}
\description{
Helper function to convert points in a single site into an sf poylgon.
}
