% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mst.R
\name{neighborsDataFrame}
\alias{neighborsDataFrame}
\title{Helper function taken from package \code{expp} on 2025-01-20 that converts an
spdep::nb neighbors object into a dataframe with columns \code{id} and \code{id_neigh}.
Function copied from \code{expp} package to avoid expp dependency as this package
is no longer maintained.}
\usage{
neighborsDataFrame(nb)
}
\description{
Helper function taken from package \code{expp} on 2025-01-20 that converts an
spdep::nb neighbors object into a dataframe with columns \code{id} and \code{id_neigh}.
Function copied from \code{expp} package to avoid expp dependency as this package
is no longer maintained.
}
