% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_pixels_to_polygons.R
\name{pixel_to_polygons}
\alias{pixel_to_polygons}
\title{Convert clustered points into collated site polygons}
\usage{
pixel_to_polygons(
  points,
  id_col,
  habitat_col,
  additional_variable_cols,
  site_id_col = "site_id",
  pixel_size = 10,
  x_col = "X",
  y_col = "Y"
)
}
\arguments{
\item{points}{data.frame. Contains a row for each pixel and cluster allocations.}

\item{id_col}{tidyselect. Column containing unique ID for the reef to be attached
to outputs.}

\item{habitat_col}{tidyselect. Column containing habitat categories.}

\item{additional_variable_cols}{tidyselect. Column(s) containing additional
continuous variables to summarise for points within sites. Output will include
site polygon median and standard deviation values for all columns selected.}

\item{site_id_col}{character or integer. Column containing site allocations for pixels.
Default = "site_id".}

\item{pixel_size}{numeric. Resolution of original habitat raster cells used to
extract point data. Must be in the same units as \code{x_col}/\code{y_col}.}

\item{x_col}{character or integer. Column containing x coordinates for points.}

\item{y_col}{character or integer. Column containing y coordinates for points.}
}
\value{
sf data.frame containing site polygons created from pixels using allocated
\code{site_id_col} values.
}
\description{
From a dataframe of pixels, this function collates points into
polygons depending on their \code{site_id_col} assignment. This method uses the
initial resolution of pixels (i.e. resolution of the habitat raster used for
extraction) to give points 2D space before being converted into polygons.
}
