% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cluster_edges.R
\name{prepare_knn_edges}
\alias{prepare_knn_edges}
\title{Create k-nearest-neighbour edges for clustering inputs.}
\usage{
prepare_knn_edges(pixels, k = 7)
}
\arguments{
\item{pixels}{sf data.frame. Holds values for pixel geometries.}

\item{k}{integer. K-nearest-neighbours}
}
\value{
Matrix containing k-n-n edge values between pixels.
}
\description{
Take a dataframe of pixels \code{pixels} and create k-nearest-neighbour
graphs between pixels based on geographic distances.
}
