% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mst.R
\name{prepare_mst}
\alias{prepare_mst}
\title{Create a minimum spanning tree from geographic coordinates of points and extracted
data.}
\usage{
prepare_mst(
  points,
  additional_variable_cols = c("depth_standard"),
  mst_alpha = 0.5,
  point_jitter = 1e-04
)
}
\arguments{
\item{points}{sf data.frame. Holds values for pixel geometries and
\code{additional_variable_cols}. Must use a projected coordinate reference system
for accurate distance calculations.}

\item{additional_variable_cols}{character vector. Names of the columns to extract
additional (non-geometric) data from for cost weighting.}

\item{mst_alpha}{float numeric. Weighting applied to \code{additional_variable_cols}
distance in edge cost weighting when combining with geographic distances.
(1 - alpha) weight is applied to the geographic distance. Default = 0.5
(same weight for \code{additional_variable_cols} and geographic distances).}

\item{point_jitter}{float numeric. Jitter applied to point coordinates for the
sole purpose of triangulation. This allows points that are in straight lines
(e.g. from raster pixel extraction) to be triangulated. A default jitter value
of 0.0001 metres is used.}
}
\value{
igraph::mst Minimum spanning tree object.
}
\description{
Take a dataframe of points \code{points} and create a minimum spanning tree
between pixel coordinates, using edge costs that are a combination of geographic
distances and distances between points by \code{additional_variable_cols} values.
For additional details on minimum spanning tree creation see igraph::mst().
}
