% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cluster_edges.R
\name{prepare_tri_edges}
\alias{prepare_tri_edges}
\title{Create triangulated edges from geographic coordinates of pixels.}
\usage{
prepare_tri_edges(pixels)
}
\arguments{
\item{pixels}{sf data.frame. Holds values for pixel geometries.}
}
\value{
Matrix containing trianguated edge values between pixels.
}
\description{
Take a dataframe of pixels \code{pixels} and triangulate the vertices.
Then extract the edges between pixels for clustering inputs.
}
