% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reef_skater.R
\name{reef_skater}
\alias{reef_skater}
\title{Cluster pixels together using the spdep::skater algorithm.}
\usage{
reef_skater(
  pixels,
  n_clust = round(min(10000, nrow(pixels))/200),
  site_size = 250 * 250,
  x_col = "X_standard",
  y_col = "Y_standard",
  habitat_col = "habitat",
  id_col = "UNIQUE_ID",
  additional_variable_cols = c("depth_standard"),
  parallelisation = "Windows",
  hex_resolution = 12
)
}
\arguments{
\item{pixels}{data.frame. Contains values for X and Y coordinates, as well as
\code{additional_variable_cols}.}

\item{n_clust}{integer numeric. Number of clusters in result output. (Point to
cut hierarchical clustering tree). Default = (round(nrow(pixels) / 200))
(dividing habitat into clusters containing an average of 200 pixels).}

\item{site_size}{numeric. Desired site size (area in m^2). Default = 625,000
(250m x 250m).}

\item{x_col}{character. Name of the column holding X coordinates. Default = "X_standard".}

\item{y_col}{character. Name of the column holding Y coordinates. Default = "Y_standard".}

\item{habitat_col}{character. Column holding unique habitat values
(attached to \code{id_col} value and site_id values on output). Default = "habitat".}

\item{id_col}{character. Column holding ID value for the target reef
(attached to the site_id values on output). Default = "UNIQUE_ID".}

\item{additional_variable_cols}{character vector. Names of additional columns
to contribute to the distance matrix. Default = c("depth_standard").}

\item{parallelisation}{character. Current option is only "Windows", using this
option sets up a parallel::Cluster using detectCores() - 2 cores. This parallelises
prunecost calculations within spdep::skater(). If \code{parallelisation} is not
set to "Windows", no parallelisation will occur. Default = "Windows".}

\item{hex_resolution}{integer numeric. H3 hexagon resolution used in pixel
creation.}
}
\value{
data.frame of pixels with allocated site_ids based on cluster outputs.
\code{site_id} values are a combination of the \code{id_col} value, \code{habitat_col} value
and the cluster allocation.
}
\description{
Take a dataframe of pixels containing geometries of pixels and
\code{additional_variable_cols} values and cluster using the skater algorithm.
This clustering is performed top-down the costs of pruning each minimum
spanning tree edge. Only pixels connected by \code{edges} are able to cluster
together. For additional information see \strong{insert citation/link}.
}
