% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefPartitionAllenAtlas-package.R
\docType{package}
\name{ReefPartitionAllenAtlas-package}
\alias{ReefPartitionAllenAtlas}
\alias{ReefPartitionAllenAtlas-package}
\title{ReefPartitionAllenAtlas: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
