% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefPartitionAllenAtlas-package.R
\docType{package}
\name{ReefPartitionUniversal-package}
\alias{ReefPartitionUniversal}
\alias{ReefPartitionUniversal-package}
\title{ReefPartitionUniversal: Partition reef habitats into sites}
\description{
Package for partitioning coral reef habitats into smaller site areas based on raster data layers. The package provides functions for extracting pixel points on a reef based on habitat types and extracting values for eadditional raster layers such as bathymetry. Points can then be clustered into sites within habitat types based on their geographic attributes as well as additional variable information such as depth.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://open-aims.github.io/ReefPartitionUniversal/}
}

}
\author{
\strong{Maintainer}: Vanessa Haller-Bull \email{V.Haller@aims.gov.au}

Authors:
\itemize{
  \item Benjamin Grier
  \item Takuya Iwanaga
}

}
\keyword{internal}
