% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_skater.R
\name{ssw}
\alias{ssw}
\title{Internal helper to compute the sum of square distances within cluster nodes.
Follows spdep function, implementing igraph methods.}
\usage{
ssw(data, nodes, method = "euclidean", p = 2, cov = NULL, inverted = FALSE)
}
\description{
Internal helper to compute the sum of square distances within cluster nodes.
Follows spdep function, implementing igraph methods.
}
