% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_raster.R
\name{vector_to_raster}
\alias{vector_to_raster}
\title{Convert vector data for a reef(s) to raster data for pixel extraction.}
\usage{
vector_to_raster(
  vector_data,
  data_column,
  output_file = NA,
  raster_template = generate_raster_template(vector_data)
)
}
\arguments{
\item{vector_data}{sf data.frame. Contains geometries and the values to convert
to raster format}

\item{data_column}{character. Column name of data to be filled into the output
raster. If the values in the column are character class, then the output
raster will contain numeric factor levels. The corresponding levels will be
output in a csv file by the name \verb{*output_file*_levels.csv}.}

\item{output_file}{character. \verb{*.tif} output file path for saving of raster data.}

\item{raster_template}{SpatRaster. Raster object to use in terra::rasterize when
converting data. By default this raster template is generated using
\code{generate_raster_template(vector_data)}, however it is recommended to use
existing raster data as a template if available.}
}
\value{
SpatRaster conversion of \code{vector_data} using a \code{raster_template}. Output
raster is also saved to \code{output_file}, and a csv containing numerical allocations
is saved if the values in \code{data_column} are characters.
}
\description{
Using terra this function converts vector data to raster data
for use in pixel extraction steps. This function is intended for use with
vector data obtained from the Allen Coral Atlas. This conversion requires
a template raster object to fill with values. If raster data is already
available, for example a continuous raster of depth values for extraction,
then it is recommended to use this raster as a template for rasterization
rather than the default generated template.
Rasterization can take a long time and a large amount of RAM if the \code{raster_template}
is fine resolution.
}
