% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_hclust.R
\name{constrained_hclust_mst}
\alias{constrained_hclust_mst}
\title{Default habitat clustering function using adespatial::constr.hclust.}
\usage{
constrained_hclust_mst(pixels, distance_alpha = 0.5, n_pixels = 204, ...)
}
\arguments{
\item{pixels}{data.frame. Contains values for X and Y coordinates, as well as
\code{additional_variable_cols}.}

\item{distance_alpha}{float numeric. Weighting applied to the additional variable
distance values when creating the distance matrix for clustering. This argument
is not included in \code{...} for discoverability.}

\item{...}{additional arguments. Additional arguments can be used here and will
be passed onto \code{prepare_mst_edges()} and \code{constrained_hclust()} functions.
These arguments must be named. \code{distance_alpha} argument is not included in
these additional arguments. For information on arguments available in these
functions and default values when arguments are not used, see \code{prepare_mst_edges()}
and \code{constrained_hclust()}.}
}
\value{
data.frame of pixels with allocated site_ids based on cluster outputs
from \code{constrained_hclust()} using \code{prepare_mst_edges} to create a minimum
spanning tree for input. \code{site_id} values are a combination of the \code{id_col}
value, \code{habitat_col} value and the cluster allocation.
}
\description{
Take a dataframe of pixels containing \code{additional_variable_cols}
values, create a minimum spanning tree using \code{prepare_mst_edges()} and then
cluster pixels using pixel data and edges with \code{constrained_hclust()}.
Any additional arguments for \code{prepare_mst_edges()} or \code{constrained_hclust()}
(excluding \code{distance_alpha}) can be included.
}
