% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexless_extraction.R
\name{fill_na_nearest}
\alias{fill_na_nearest}
\title{Internal helper function that fills NA values from columns with values from
the nearest neighbouring points.}
\usage{
fill_na_nearest(pixel_data, columns)
}
\description{
Internal helper function that fills NA values from columns with values from
the nearest neighbouring points.
}
