% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cluster_edges.R
\name{prepare_mst}
\alias{prepare_mst}
\title{Create a minimum spanning tree from geographic coordinates of pixels and extracted
data.}
\usage{
prepare_mst(
  pixels,
  additional_variable_cols = c("depth_standard"),
  alpha = 0.5,
  hex_resolution = 12
)
}
\arguments{
\item{pixels}{sf data.frame. Holds values for pixel geometries and
\code{additional_variable_cols}}

\item{additional_variable_cols}{character vector. Names of the columns to extract
additional (non-geometric) data from for cost weighting.}

\item{alpha}{float numeric. Weighting applied to \code{additional_variable_cols}
distance in edge cost weighting when combining with geographic distances.
(1 - alpha) weight is applied to the geographic distance. Default = 0.5
(same weight for \code{additional_variable_cols} and geographic distances).}

\item{hex_resolution}{integer. H3 hex resolution of pixels. Default = 12.}
}
\value{
igraph::mst Minimum spanning tree object.
}
\description{
Take a dataframe of pixels \code{pixels} and create a minimum spanning tree
between pixel coordinates, using edge costs that are a combination of geographic
distances and distances between pixels by \code{additional_variable_cols} values.
For additional details on minimum spanning tree creation see igraph::mst().
}
