% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vh_skater.R
\name{skater_igraph}
\alias{skater_igraph}
\title{An approximate Skater algorithm implemented using igraph methods.}
\usage{
skater_igraph(
  edges,
  data,
  ncuts,
  crit,
  vec.crit,
  method = "euclidean",
  p = 2,
  cov = NULL,
  inverted = FALSE
)
}
\description{
Leverages C implementations in igraph package.
The approach collects all edges from all clusters, sorts by global cost, and
iterates until a valid split is found.
}
